Correct installation guide for *nix
When specifying alias, there shouldn't be any empty space in between alias name and equal to sign. 
i.e. `alias njexl = ..` must be `alias njexl=...`.
Else it is read as a character.

Also added doc to reload the login file (makes it effortless for people who are not aware of it)